requirements.txt
# Core ML/Deep Learning
# CPU-only install:  pip install -r requirements.txt
# GPU install:       pip install -r requirements-gpu.txt && pip install -r requirements.txt
#                    (requirements-gpu.txt installs the CUDA torch + nvidia-ml-py first;
#                     pip will then skip torch here because >=2.5.1 is already satisfied)
torch>=2.5.1
torchvision>=0.20.1
transformers==4.45.2
accelerate==1.5.1

# Data processing
numpy==1.26.4
pandas==2.2.3
pillow==11.1.0
opencv-python==4.11.0.86
scikit-learn==1.6.1
pyarrow==19.0.0

# Training utilities
evaluate==0.4.3
jiwer==4.0.0
tensorboard==2.8.0
tqdm==4.67.1

# Comparison feature
python-Levenshtein>=0.23.0  # Fast edit distance for CER/WER calculation

# Configuration
pyyaml==6.0.2

# GPU monitoring (optional — in requirements-gpu.txt; code gracefully skips if absent)
# nvidia-ml-py>=12.0.0

# Web UI (polyscriptor_server.py — browser-based HTR interface)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9  # multipart/form-data file uploads
pymupdf>=1.24.0           # PDF → image rendering

# Notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.29.0
matplotlib==3.10.1
